Refactor dropdown to selectbox
