+++ components/Selectbox/README.md
@@ -1,10 +1,10 @@
-# Dropdown
+# Selectbox
 
-The dropdown component renders a simple dropdown based on user defined options.
+The Selectbox component renders a simple Selectbox based on user defined options.
 
 ```jsx static
 const options = [{ value: 'A', label: 'A' }, { value: 'B', label: 'B' }];
 const handleChange = (e: HTMLInputElement) => alert('changed to value: ' + e.value);
 
-<Dropdown options={options} value={options[0].value} onChange={handleChange} />;
+<Selectbox options={options} value={options[0].value} onChange={handleChange} />;
 ```
