+++ components/Selectbox/Selectbox.stories.tsx
@@ -1,14 +1,14 @@
 import * as React from 'react';
 import { storiesOf } from '@storybook/react';
-import { Dropdown } from './Dropdown';
+import { Selectbox } from './Selectbox';
 
 const options = [{ value: 'A', label: 'A' }, { value: 'B', label: 'B' }];
 
-storiesOf('Dropdown', module)
-    .add('should display a dropdown', () => {
-        return <Dropdown options={options} value={options[0].value} />;
+storiesOf('Selectbox', module)
+    .add('should display a selectbox', () => {
+        return <Selectbox options={options} value={options[0].value} />;
     })
     .add('should alert on change', () => {
         const handleChange = (e: HTMLInputElement) => alert('changed to value: ' + e.value);
-        return <Dropdown options={options} value={options[0].value} onChange={handleChange} />;
+        return <Selectbox options={options} value={options[0].value} onChange={handleChange} />;
     });
